fix dtok

+++ database/migrations/2021_03_05_195949_buat_produk_table.php
@@ -21,7 +21,7 @@ class BuatProdukTable extends Migration
             $table->integer('harga_beli');
             $table->tinyInteger('diskon')->default(0);
             $table->integer('harga_jual');
-            $table->integer('stok');
+            $table->float('stok');
             $table->timestamps();
         });
     }
